refactor: 타입 패키지 이동
