+++ android/app/src/main/java/com/now/naaga/data/mapper/AuthMapper.kt
@@ -1,7 +1,7 @@
 package com.now.naaga.data.mapper
 
-import com.now.domain.model.AuthPlatformType
 import com.now.domain.model.PlatformAuth
+import com.now.domain.model.type.AuthPlatformType
 import com.now.naaga.data.remote.dto.PlatformAuthDto
 
 fun PlatformAuthDto.toDomain(): PlatformAuth {
