+++ android/app/src/main/java/com/now/naaga/presentation/login/LoginActivity.kt
@@ -8,7 +8,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import com.now.domain.model.AuthPlatformType
+import com.now.domain.model.type.AuthPlatformType
 import com.now.naaga.data.firebase.analytics.AnalyticsDelegate
 import com.now.naaga.data.firebase.analytics.DefaultAnalyticsDelegate
 import com.now.naaga.data.firebase.analytics.LOGIN_AUTH
