+++ android/domain/src/main/java/com/now/domain/model/PlatformAuth.kt
@@ -1,5 +1,7 @@
 package com.now.domain.model
 
+import com.now.domain.model.type.AuthPlatformType
+
 data class PlatformAuth(
     val token: String,
     val type: AuthPlatformType,
