+++ android/domain/src/main/java/com/now/domain/model/type/OrderType.kt
@@ -1,4 +1,4 @@
-package com.now.domain.model
+package com.now.domain.model.type
 
 enum class OrderType {
     ASCENDING,
